+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.1-RC16]
+
+- LeiaMessageProduceClient
+  - MessageProcessorHub now returns a list of processors.
+
+
 ## [0.0.1-RC15]
 
 - LeiaMessageProduceClient
-  - Started taking a messageProcessorHub in the client. A subset of messages could be pawned to each processor
+    - Started taking a messageProcessorHub in the client. A subset of messages could be pawned to each processor
 
 ## [0.0.1-RC14]